+++ lib/src/helpers.dart
@@ -9,7 +9,3 @@ bool isCompact(context) => deviceWidth(context) < Styles.compact;
 bool isMedium(context) => !isCompact(context) && !isExpanded(context);
 
 bool isExpanded(context) => deviceWidth(context) > Styles.expanded;
-
-BorderRadius borderRadiusAll(double borderRadius) => BorderRadius.all(
-      Radius.circular(borderRadius),
-    );